requirements.txt
# Core ML
torch>=2.6.0
torchvision
torchaudio
transformers>=4.36.0
safetensors
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.2.0

# Graph Neural Network
torch-geometric>=2.4.0

# AST Parsing
tree-sitter>=0.20.0

# Explainability
shap>=0.43.0

# YARA
yara-python>=4.3.0

# LLM
openai>=1.6.0
anthropic>=0.8.0

# Frontend
streamlit>=1.29.0

# Visualization
matplotlib>=3.7.0
plotly>=5.18.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
requests>=2.31.0